@assert/primitives initialization
The introductory commit presenting a JavaScript assertion library. This
project is going to be revolutionary (lol) and will be used by
generations to come! It will be studied in universities and by aliens
that attempt to understand our glorious intelligence. There will be no
greater source of intellect that can provide as much insight into the
human congition than this piece of work.